Refactor /introspect endpoint
