+++ web/app/api/v1/oidc/introspect/route.ts
@@ -1,0 +1,1 @@
+export { POST } from "@/api/v1/oidc/introspect";
